pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "archdyn"
version = "0.1.0"
description = "Minimal research codebase for CINIC-10 CNN vs ViT experiments."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "matplotlib>=3.8",
  "numpy>=1.26",
  "pandas>=2.2",
  "PyYAML>=6.0",
  "scikit-learn>=1.4",
  "timm>=1.0.7",
  "torch>=2.2",
  "torchvision>=0.17",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
